Refactor components to use same structure
