+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/hasan/Downloads/microverse/bookstore/src/components/CategoryFilter.js":"1"},{"size":525,"mtime":1607088651063,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"116wozx","/home/hasan/Downloads/microverse/bookstore/src/components/CategoryFilter.js",[]]
+[{"/home/hasan/Downloads/microverse/bookstore/src/components/BookStoreHeader.js":"1","/home/hasan/Downloads/microverse/bookstore/src/components/UserForm.js":"2","/home/hasan/Downloads/microverse/bookstore/src/components/Navbar.js":"3","/home/hasan/Downloads/microverse/bookstore/src/components/Loading.js":"4","/home/hasan/Downloads/microverse/bookstore/src/components/CategoryFilter.js":"5","/home/hasan/Downloads/microverse/bookstore/src/components/Book.js":"6","/home/hasan/Downloads/microverse/bookstore/src/components/App.js":"7","/home/hasan/Downloads/microverse/bookstore/src/container/BooksForm.js":"8","/home/hasan/Downloads/microverse/bookstore/src/container/BooksList.js":"9","/home/hasan/Downloads/microverse/bookstore/src/container/CategoryFilterContainer.js":"10","/home/hasan/Downloads/microverse/bookstore/src/container/NavbarContainer.js":"11","/home/hasan/Downloads/microverse/bookstore/src/container/SignUp.js":"12","/home/hasan/Downloads/microverse/bookstore/src/container/UserFormContainer.js":"13"},{"size":352,"mtime":1607091100449,"results":"14","hashOfConfig":"15"},{"size":837,"mtime":1607090944995,"results":"16","hashOfConfig":"15"},{"size":969,"mtime":1607091076361,"results":"17","hashOfConfig":"15"},{"size":87,"mtime":1607090893627,"results":"18","hashOfConfig":"15"},{"size":472,"mtime":1607091089689,"results":"19","hashOfConfig":"15"},{"size":1965,"mtime":1607091111869,"results":"20","hashOfConfig":"15"},{"size":906,"mtime":1607091124889,"results":"21","hashOfConfig":"15"},{"size":2563,"mtime":1607090953675,"results":"22","hashOfConfig":"15"},{"size":1026,"mtime":1607091055144,"results":"23","hashOfConfig":"15"},{"size":578,"mtime":1607091040128,"results":"24","hashOfConfig":"15"},{"size":494,"mtime":1607091015072,"results":"25","hashOfConfig":"15"},{"size":449,"mtime":1607090979640,"results":"26","hashOfConfig":"15"},{"size":1055,"mtime":1607090988652,"results":"27","hashOfConfig":"15"},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"116wozx",{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"42","messages":"43","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"44","messages":"45","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"46","messages":"47","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"48","messages":"49","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"50","messages":"51","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"52","messages":"53","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/hasan/Downloads/microverse/bookstore/src/components/BookStoreHeader.js",[],"/home/hasan/Downloads/microverse/bookstore/src/components/UserForm.js",[],"/home/hasan/Downloads/microverse/bookstore/src/components/Navbar.js",[],"/home/hasan/Downloads/microverse/bookstore/src/components/Loading.js",[],"/home/hasan/Downloads/microverse/bookstore/src/components/CategoryFilter.js",[],"/home/hasan/Downloads/microverse/bookstore/src/components/Book.js",[],"/home/hasan/Downloads/microverse/bookstore/src/components/App.js",[],"/home/hasan/Downloads/microverse/bookstore/src/container/BooksForm.js",[],"/home/hasan/Downloads/microverse/bookstore/src/container/BooksList.js",[],"/home/hasan/Downloads/microverse/bookstore/src/container/CategoryFilterContainer.js",[],"/home/hasan/Downloads/microverse/bookstore/src/container/NavbarContainer.js",[],"/home/hasan/Downloads/microverse/bookstore/src/container/SignUp.js",[],"/home/hasan/Downloads/microverse/bookstore/src/container/UserFormContainer.js",[]]
