+++ src/components/CategoryFilter.js
@@ -1,17 +1,14 @@
-import React from 'react';
 import PropTypes from 'prop-types';
 
-function CategoryFilter({ handleFilterChange, categories }) {
-  return (
-    <select onChange={handleFilterChange} name="filter">
-      {categories.map(category => (
-        <option key={category} value={category}>
-          {category}
-        </option>
-      ))}
-    </select>
-  );
-}
+const CategoryFilter = ({ handleFilterChange, categories }) => (
+  <select onChange={handleFilterChange} name="filter">
+    {categories.map(category => (
+      <option key={category} value={category}>
+        {category}
+      </option>
+    ))}
+  </select>
+);
 
 CategoryFilter.propTypes = {
   handleFilterChange: PropTypes.func.isRequired,
