+++ src/components/Loading.js
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const Loading = () => (
   <div className="loading main" />
 );
