+++ src/components/UserForm.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import PropTypes from 'prop-types';
 import Loading from './Loading';
