+++ src/container/Login.js
@@ -1,9 +1,8 @@
-import React from 'react';
 import { Redirect } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 
 import { fetchUserLogin } from '../redux';
-import UserFormContainer from '../container/UserFormContainer';
+import UserFormContainer from './UserFormContainer';
 
 const Login = () => {
   const user = useSelector(state => state.user);
