+++ src/container/SignUp.js
@@ -1,9 +1,8 @@
-import React from 'react';
 import { Redirect } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 
 import { fetchUserSignUp } from '../redux';
-import UserFormContainer from '../container/UserFormContainer';
+import UserFormContainer from './UserFormContainer';
 
 const SignUp = () => {
   const user = useSelector(state => state.user);
